Move original tests to separate class
